Fix the GC branch so it compiles and runs on 64 bit. Incremental marking is switched off for the moment. Review URL: http://codereview.chromium.org/7015043
git-svn-id: http://v8.googlecode.com/svn/branches/experimental@7904 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ gc/src/incremental-marking.h
@@ -233,7 +233,7 @@ class IncrementalMarking : public AllStatic {
   int steps_count_;
   double steps_took_;
   bool should_hurry_;
-  intptr_t allocation_marking_factor_;
+  int allocation_marking_factor_;
   intptr_t allocated_;
 };
